requirements.txt
crewai 
crewai-tools 
chonkie[semantic] 
markitdown 
qdrant-client 
fastembed
streamlit
python-dotenv